Refactor clasp pull test
